2005-12-19 Roman - !!!!:/Parameter: Made parameters logical level. Created parameter's templates and logical level parameters. - /Doc: Add umbrello uml diagram to repository. - /Parameter: Add base error info atributes: "err" and "err_mess". - !!!:/Parameter: Separate parameter name to Id, Name and Description. - !!:/Parameter: Remove export concept and use logical level for it. - /Value: Change mirroring config concept. Add set config function TValue::vlCfg(). - !!!:/Function: Add blocking change concept for the function values. - /BD: Add the interface SQL request function virtual TBD::sqlReq( ). - !!!:/BD: Move open DB procedures to functions: dataSeek(),dataGet() and dataSet(). - !!:/BD: Add generic data delete function TBD::dataDel(). - !:/System: Adapted to GCC compiler version 4.0.1. - /Mess: Add checking to equality the coding for translate messages. - /System: Change the creating and starting subsystem order. - !!!:/BD/SQLite: Add the set encoding support. DB encoding set into DB name: <./my.db;KOI8-R>. - !!!:/BD/MySQL: Add the set encoding support. DB encoding set into DB name:     <server.diya.org;root;;oscadaDemoSt;;;KOI8-U>. - !!!:/BD/DBF: Add the set encoding support. DB encoding set into Table name: <./my.dbf;KOI8-R>. - /Controller/OperationSystem: Add the autocreating option to the parameter for not saving it. - /Controller/OperationSystem: Add translated names of autocreating parameters. - /Controller/JavaLikeCalc: Fix UTF8 symbols check. - /Controller/BlockCalc: Change functions: void loadIO( ) and void saveIO( ).     Remove the hot function structure changing support! - /Controller/BlockCalc: Add the error messages generation. - /UI/QTCfg: Add the self table's columns permison support. - /UI/QTCfg: Add id fields Tool Tips. - /UI/QTCfg: Add select mode "sel_ed" with edit support. - /UI/Vision: Fix many warnings. - /UI/WebCfg: Add the self table's columns permison support. - /UI/WebCfg: Add select mode "sel_ed" with edit support.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@461 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tsys.h
@@ -116,7 +116,7 @@ class TSYS : public TCntrNode
 	}
         unsigned long long shrtCnt( )
         {
-    	    unsigned long cntl, cnth;
+    	    unsigned int cntl, cnth;
     	    asm volatile("rdtsc; movl %%eax,%0; movl %%edx,%1;":"=r"(cntl),"=r"(cnth)::"%eax","%edx");
 	    return ((unsigned long long)cnth<<32)+cntl;	
         }
